ConversionTable stub.
This module should do the conversion of contexts from string based
representations as found in the corpus, to index-based representations
(given a conversion table) as found in the context database.

As it is, I seem to lack parameterization. Specifically, one should be
able to add the connection *and* table within the connection in order to
handle lemma lookups.