Completely rewrite of recordContext.
I feel like I just gave birth to an elephant.

This is a substantial rewrite, accomodating the new format and it also
has a nice multi-threaded logging capability that will keep us up to
date on progress.

This is the untested version. There's no way this thing will actually
run, it needs some debugging first.