Remove unigram indexing.
Alright, premature optimization is the root of all evil. Again.
I removed all the work on the unigram tables — we now insert plain text
values into the database, not integer indices.
Anwyay, the training process (and probably matching, too) will be *much*
faster this way.
Maybe it'll take up some more space in the end, but space is cheap, time
is not.