Make ItemGetters work with Documents.
This makes corpusI obsolete.

Remove corpusI.

I had the idea of making an Iteratee (Document Text) m (Sentence Text),
but that turned out harder than expected (the general case would have
been (Nullable a) => Iteratee [a] m a, or somesuch, but I don't think
that's really easy. Mabye a question for SO?