Use periods as sentence markers.
That means, insert them at the beginning of each document, also insert
them at the beginning of all sentences that are later to serve as items.

Add gzip decompression functionality.

Revert "Add gzip decompression functionality."

This reverts commit 16a38133678bb5c95771db66bf8ea3247fef9814. It seems
that this won't work because of how zlib is coded. Maybe that can be
changed upstream, but there's nothing we can do about it here. We'll
have to leave this (very practical!) functionality out for now…